fix issue

+++ quick_item.py
@@ -40,7 +40,7 @@ def display_session(id):
     session = db.get_by_session_id(id)
     hours = db.get_user_pref(st.session_state["username"])["preferred_hours_per_session"]
 
-    increment = 10 * hours
+    increment = int(10 * hours)
     with st.container(border = True):
         col1,col2 = st.columns([1,9])
         with col1:
